backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
anthropic>=0.39.0
chromadb>=0.6.0
sentence-transformers>=3.0.0
pydantic>=2.9.0
python-multipart>=0.0.9
unstructured>=0.15.0
celery>=5.4.0
redis>=5.0.0
aiosqlite>=0.20.0
diff-match-patch>=20230430
openpyxl>=3.1.5
python-dotenv>=1.0.1
openai>=1.0.0
requests>=2.31.0
bcrypt>=4.1.0
PyJWT>=2.8.0
slowapi>=0.1.9
pytest>=8.0.0
httpx>=0.27.0
